+++ src/Graph/Graphs/Product.php
@@ -22,11 +22,23 @@ class Product extends AbstractGraph
 {
             id
             title
+            handle
             totalInventory
             tracksInventory
             status
             productType
             description
+            descriptionHtml
+            onlineStoreUrl
+            vendor
+            status
+            isGiftCard
+            requiresSellingPlan
+            tags
+            seo {
+                title
+                description
+            }
             images(first: 250) {
                 edges {
                     node {
@@ -34,19 +46,75 @@ class Product extends AbstractGraph
                     }
                 }
             }
+            media(first: 250) {
+                edges {
+                    node {
+                        id
+                        mediaContentType
+                    }
+                }
+            }
             category {
+                id
                 name
             }
-            status
             priceRangeV2 {
                 minVariantPrice {
                     amount
+                    currencyCode
                 }
                 maxVariantPrice {
                     amount
+                    currencyCode
+                }
+            }
+            collections(first: 250) {
+                edges {
+                    node {
+                        id
+                        title
+                        description
+                        image {
+                            url
+                        }
+                    }
+                }
+            }
+            variants(first: 250) {
+                edges {
+                    node {
+                        id
+                        title
+                        barcode
+                        sku
+                    }
+                }
+            }
+            options(first: 250) {
+                id
+                name
+                optionValues {
+                    id
+                    name
+                    swatch {
+                        color
+                        image {
+                            id
+                        }
+                    }
+                }
+            }
+            sellingPlanGroups(first: 250) {
+                edges {
+                    node {
+                        name
+                    }
                 }
             }
             $extraGraph
+            createdAt
+            updatedAt
+            publishedAt
         }
 GRAPH;
     }