+++ CHANGELOG.md
@@ -16,3 +16,7 @@
 
 - Add product count
 - Expand full product
+
+## [1.0.4] - 2026-03-25
+
+- Add catalog endpoint